+++ src/lib.rs
@@ -267,6 +267,7 @@ pub mod traits;
 pub mod types;
 mod wide_div;
 mod wrapping;
+
 use crate::{
     arith::MulDivOverflow,
     from_str::FromStrRadix,
@@ -312,20 +313,22 @@ mod macros_frac;
 macro_rules! fixed {
     (
         $description:expr,
-        $Fixed:ident($Inner:ty, $LeEqU:tt, $s_nbits:expr),
+        $Fixed:ident($Inner:ty, $LeEqU:tt, $s_nbits:expr, $s_nbits_m4:expr),
         $nbytes:expr, $bytes_val:expr, $be_bytes:expr, $le_bytes:expr,
         $UInner:ty, $Signedness:tt
     ) => {
         fixed! {
             $description,
-            $Fixed[stringify!($Fixed)]($Inner[stringify!($Inner)], $LeEqU, $s_nbits),
+            $Fixed[stringify!($Fixed)]($Inner[stringify!($Inner)], $LeEqU, $s_nbits, $s_nbits_m4),
             $nbytes, $bytes_val, $be_bytes, $le_bytes,
             $UInner, $Signedness
         }
     };
     (
         $description:expr,
-        $Fixed:ident[$s_fixed:expr]($Inner:ty[$s_inner:expr], $LeEqU:tt, $s_nbits:expr),
+        $Fixed:ident[$s_fixed:expr](
+            $Inner:ty[$s_inner:expr], $LeEqU:tt, $s_nbits:expr, $s_nbits_m4:expr
+        ),
         $nbytes:expr, $bytes_val:expr, $be_bytes:expr, $le_bytes:expr,
         $UInner:ty, $Signedness:tt
     ) => {
@@ -401,7 +404,7 @@ assert_eq!(two_point_75.to_string(), \"2.8\");
         // inherent methods that require Frac bounds, and cannot be const
         fixed_frac! {
             $description,
-            $Fixed[$s_fixed]($Inner[$s_inner], $LeEqU, $s_nbits),
+            $Fixed[$s_fixed]($Inner[$s_inner], $LeEqU, $s_nbits, $s_nbits_m4),
             $UInner, $Signedness
         }
     };
@@ -409,25 +412,25 @@ assert_eq!(two_point_75.to_string(), \"2.8\");
 
 fixed! {
     "An eight-bit fixed-point unsigned",
-    FixedU8(u8, LeEqU8, "8"),
+    FixedU8(u8, LeEqU8, "8", "4"),
     1, "0x12", "[0x12]", "[0x12]",
     u8, Unsigned
 }
 fixed! {
     "A 16-bit fixed-point unsigned",
-    FixedU16(u16, LeEqU16, "16"),
+    FixedU16(u16, LeEqU16, "16", "12"),
     2, "0x1234", "[0x12, 0x34]", "[0x34, 0x12]",
     u16, Unsigned
 }
 fixed! {
     "A 32-bit fixed-point unsigned",
-    FixedU32(u32, LeEqU32, "32"),
+    FixedU32(u32, LeEqU32, "32", "28"),
     4, "0x1234_5678", "[0x12, 0x34, 0x56, 0x78]", "[0x78, 0x56, 0x34, 0x12]",
     u32, Unsigned
 }
 fixed! {
     "A 64-bit fixed-point unsigned",
-    FixedU64(u64, LeEqU64, "64"),
+    FixedU64(u64, LeEqU64, "64", "60"),
     8, "0x1234_5678_9ABC_DEF0",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0xF0]",
     "[0xF0, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
@@ -435,7 +438,7 @@ fixed! {
 }
 fixed! {
     "A 128-bit fixed-point unsigned",
-    FixedU128(u128, LeEqU128, "128"),
+    FixedU128(u128, LeEqU128, "128", "124"),
     16, "0x1234_5678_9ABC_DEF0_1234_5678_9ABC_DEF0",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0xF0, \
      0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0xF0]",
@@ -445,25 +448,25 @@ fixed! {
 }
 fixed! {
     "An eight-bit fixed-point signed",
-    FixedI8(i8, LeEqU8, "8"),
+    FixedI8(i8, LeEqU8, "8", "4"),
     1, "0x12", "[0x12]", "[0x12]",
     u8, Signed
 }
 fixed! {
     "A 16-bit fixed-point signed",
-    FixedI16(i16, LeEqU16, "16"),
+    FixedI16(i16, LeEqU16, "16", "12"),
     2, "0x1234", "[0x12, 0x34]", "[0x34, 0x12]",
     u16, Signed
 }
 fixed! {
     "A 32-bit fixed-point signed",
-    FixedI32(i32, LeEqU32, "32"),
+    FixedI32(i32, LeEqU32, "32", "28"),
     4, "0x1234_5678", "[0x12, 0x34, 0x56, 0x78]", "[0x78, 0x56, 0x34, 0x12]",
     u32, Signed
 }
 fixed! {
     "A 64-bit fixed-point signed",
-    FixedI64(i64, LeEqU64, "64"),
+    FixedI64(i64, LeEqU64, "64", "60"),
     8, "0x1234_5678_9ABC_DEF0",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0xF0]",
     "[0xF0, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
@@ -471,7 +474,7 @@ fixed! {
 }
 fixed! {
     "A 128-bit fixed-point signed",
-    FixedI128(i128, LeEqU128, "128"),
+    FixedI128(i128, LeEqU128, "128", "124"),
     16, "0x1234_5678_9ABC_DEF0_1234_5678_9ABC_DEF0",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0xF0, \
      0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0xF0]",